fix: adjusted imageRes width values
