+++ src/lib/useYoutubeThumbnail.tsx
@@ -1,19 +1,41 @@
 import { useState, useEffect } from "react";
 
+export type imgResolution =
+  | "default"
+  | "mqdefault"
+  | "hqdefault"
+  | "sddefault"
+  | "maxresdefault";
+
+const expectedWidths: Record<imgResolution, number> = {
+  default: 120,
+  mqdefault: 320,
+  hqdefault: 480,
+  sddefault: 640,
+  maxresdefault: 1280,
+};
+
 export const useYoutubeThumbnail = (
   videoId: string,
   vi: string,
-  format: string
+  format: string,
+  imageRes: imgResolution = "maxresdefault"
 ) => {
   const [url, setUrl] = useState("");
 
   useEffect(() => {
-    const testUrl = `https://img.youtube.com/${vi}/${videoId}/maxresdefault.${format}`;
+    const testUrl = `https://img.youtube.com/${vi}/${videoId}/${imageRes}.${format}`;
     const fallbackUrl = `https://img.youtube.com/${vi}/${videoId}/hqdefault.${format}`;
+
+    const expectedWidth = expectedWidths[imageRes];
 
     const img = new Image();
     img.onload = () => {
-      setUrl(img.width > 120 ? testUrl : fallbackUrl);
+      if (img.width < expectedWidth) {
+        setUrl(fallbackUrl);
+      } else {
+        setUrl(testUrl);
+      }
     };
     img.onerror = () => setUrl(fallbackUrl);
     img.src = testUrl;
